*     iiblnd fixes (mid-way through changing Infinicon API)
*     viblnd fix (tx_waiting not cleared on RDMA ops initiated on a new
      connection that fails triggers an assertion failure).

*     some prep for userspace ip2nets

*     router selection round robins if other selection criterea are equal

*     local_nid_dist_zero LNET module param for single-node LND testing.

*     reformat LNET /proc buffer displays

*     rename userspace tcplnd env params TCPNAL_xxx -> TCPLND_xxx

+++ lnet/utils/portals.c
@@ -1260,7 +1260,7 @@ lwt_snapshot(cycles_t *now, int *ncpu, int *totalsize,
         if (data.ioc_u32[2] != sizeof(lwt_event_t) ||
             data.ioc_u32[3] != offsetof(lwt_event_t, lwte_where)) {
                 fprintf(stderr,"kernel/user LWT event mismatch %d(%d),%d(%d)\n",
-                        (int)data.ioc_u32[2], sizeof(lwt_event_t),
+                        (int)data.ioc_u32[2], (int)sizeof(lwt_event_t),
                         (int)data.ioc_u32[3],
                         (int)offsetof(lwt_event_t, lwte_where));
                 return (-1);
